requirements.txt
# Core data science libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Statistical analysis
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Interactive Brokers API
ib_insync>=0.9.85

# Real-time data
yfinance>=0.2.18

# Inter-process communication
pyzmq>=25.0.0

# Configuration management
pyyaml>=6.0

# Logging and monitoring
colorlog>=6.7.0
psutil>=5.9.0

# Kalman Filter implementation
pykalman>=0.0.1

# Financial analysis
arch>=6.2.0

# Plotting (for analysis)
matplotlib>=3.7.0
plotly>=5.15.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
